% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPlot.R
\name{AnnotationPlot}
\alias{AnnotationPlot}
\alias{AnnotationPlot.Seurat}
\alias{AnnotationPlot.SingleCellExperiment}
\alias{AnnotationPlot.DimReduc}
\alias{AnnotationPlot.nmf}
\alias{AnnotationPlot.list}
\alias{AnnotationPlot.data.frame}
\title{Plot annotations from an NMF model or other compatible objects.}
\usage{
AnnotationPlot(object, ...)

\method{AnnotationPlot}{Seurat}(
  object,
  plot.field = NULL,
  reduction = "nmf",
  dropEmpty = TRUE,
  annotation.name = "annotations",
  cluster_by = "pos",
  ...
)

\method{AnnotationPlot}{SingleCellExperiment}(
  object,
  plot.field = NULL,
  reduction = "nmf",
  dropEmpty = TRUE,
  annotation.name = "annotations",
  cluster_by = "pos",
  ...
)

\method{AnnotationPlot}{DimReduc}(
  object,
  plot.field = NULL,
  dropEmpty = TRUE,
  annotation.name = "annotations",
  cluster_by = "pos",
  ...
)

\method{AnnotationPlot}{nmf}(
  object,
  plot.field = NULL,
  dropEmpty = TRUE,
  annotation.name = "annotations",
  cluster_by = "pos",
  ...
)

\method{AnnotationPlot}{list}(object, plot.field, dropEmpty = TRUE, cluster_by = "pos", ...)

\method{AnnotationPlot}{data.frame}(object, plot.field, dropEmpty = TRUE, cluster_by = "pos", ...)
}
\arguments{
\item{object}{a compatible object (Seurat, DimReduc, nmf, data.frame)}

\item{...}{additional arguments passed to called functions}

\item{plot.field}{metadata grouping to plot}

\item{reduction}{the reduction to plot (default is 'nmf')}

\item{dropEmpty}{drop factors without significant associations? (TRUE)}

\item{annotation.name}{name of factor annotation, default = "annotations"}
}
\value{
a ggplot2 object
}
\description{
After running \code{AnnotateNMF}, this function returns 
a dot plot of the results

After running \code{AnnotateNMF}, this function returns 
a dot plot of the results.  Right now the code is the same as for DimReduc.
}
\examples{
\dontrun{
get_pbmc3k_data() \%>\% NormalizeData \%>\% RunNMF \%>\% AnnotateNMF -> pbmc3k
AnnotationPlot(pbmc3k, "cell_type")
}
\dontrun{
dat <- pbmc3k@reductions$nmf@misc$annotations$cell_type
AnnotationPlot(dat, "cell_type")

# if running interactively:
library(plotly)
ggplotly(AnnotationPlot(dat, "cell_type"))
}
}
