% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoNMF.R
\name{AutoNMF}
\alias{AutoNMF}
\title{Automatic Non-negative Matrix Factorization}
\usage{
AutoNMF(
  object,
  reps = 3,
  tol = 1e-05,
  maxit = 100,
  verbose = 2,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  use.seed = NULL,
  ...
)
}
\arguments{
\item{object}{a dense or sparse matrix to factorize}

\item{reps}{number of times to repeat the automatic rank determination procedure before learning the final model once on the consensus best rank}

\item{tol}{stopping criteria for the final model fit. Given as the Pearson correlation distance between \code{w} across consecutive iterations.}

\item{maxit}{stopping criteria for the final model fit. Maximum number of alternating least squares iterations.}

\item{L1}{L1/LASSO regularization penalty in the range (0, 1] to increase the sparsity of each factor in the model.}

\item{L2}{L2/Ridge regularization penalty in the range (0, 1] to increase the angle between factors in the model.}

\item{threads}{number of threads to use.}

\item{...}{advanced parameters with reasonable defaults that should not normally need to be changed:
\code{auto.subsample.exponent = 10},
\code{auto.k.init = 2},
\code{auto.step.size = 2},
\code{auto.test.set.density = 0.05},
\code{auto.maxit = 100},
\code{auto.overfit.tol = 1e-3},
\code{auto.tol = 1e-4}}
}
\description{
AutoNMF quickly learns an NMF model on a matrix at the largest rank for which overfitting does not occur, which usually is the rank that maximizes imputation accuracy and generalization of the model to other datasets.
In other words, AutoNMF learns a maximally robust and minimally biased model.
}
\details{
AutoNMF uses a combination of downsampling, coordinate descent, and detection of overfitting to quickly determine the optimal rank of an NMF model. This automatic rank determination procedure is the following:
\enumerate{
\item Select a random sample of columns, where the sample size is a power of 2 given by \code{auto.subsample.exponent = 10} by default.  For example, \code{2^10} corresponds to 1024 samples.
\item Select an initial rank (default \code{auto.k.init = 2}) and a step size (default \code{auto.step.size = 2}). Using this algorithm, it will be impossible to decrease the rank from the initial value provided.
\item Initialize \code{w} with values in a random uniform distribution.  Define a test set as a random speckled pattern with some density (default \code{auto.test.set.density = 0.05}).
\item Fit the model at the initial rank, masking test set values. Stop alternating least squares updates when any of three stopping criterion are satisfied:
\enumerate{
\item a maximum number of iterations (default \code{auto.maxit = 100})
\item Pearson correlation distance between models across consecutive iterations (default \code{auto.tol = 1e-4})
\item significant increase in test set reconstruction error over the minimum test set reconstruction error during model fitting (default \code{auto.overfit.tol = 1e-3})
}
\item If the model fitting was not terminated due to overfitting, record the test set reconstruction error at convergence. Increase the rank by \code{auto.step.size}, refit, and repeat this procedure until overfitting is detected.
\item If the model is overfit, increase the column sample size by a power of 2 and refit.
\item Once the sample size cannot be increased by a power of 2 without exceeding the number of samples in the dataset, reduce the rank by \code{auto.step.size} from the last detected overfit rank, remove the test set, and factorize the full dataset.
}

The seed may be specified directly to the function using \code{use.seed}. If \code{use.seed = NULL}, the seed will be derived from the current R environment \code{.Random.seed}. The random seed will control downsampling, test set membership, and initialization of the model.
}
