% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellxgene_pipeline.R
\name{cellxgene_pipeline}
\alias{cellxgene_pipeline}
\title{Learn an NMF model from a cellxgene Seurat object}
\usage{
cellxgene_pipeline(url, reps = 1, verbose = 3, L1 = 0.05, ...)
}
\arguments{
\item{url}{download url for a Seurat v4 object}

\item{...}{arguments to \code{RunNMF}}
}
\description{
Provide a link to download a cellxgene Seurat object, and this pipeline will return a standardized annotated NMF object at the optimal rank
}
\details{
This pipeline runs the following steps:
\enumerate{
\item Download a Seurat v4 object from the provided URL
\item Preprocess the data and run NMF using parameters specified in the \code{...} argument
\item Annotate the NMF model against existing multi-level factor information
\item Extract the model and annotations and save to an RDS file
}
}
