% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColumns.R
\name{checkColumns}
\alias{checkColumns}
\title{verify that columns for auto-annotation are factors with > 1 level}
\usage{
checkColumns(meta.data, columns = NULL)
}
\arguments{
\item{meta.data}{the meta.data (or a Seurat object if needs be)}

\item{columns}{the columns (optional; if NULL, will check all columns)}
}
\value{
a vector of suitable columns (may be length 0)
}
\description{
verify that columns for auto-annotation are factors with > 1 level
}
