% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDesigns.R
\name{checkDesigns}
\alias{checkDesigns}
\title{verify that a list of matrices is in fact a named list of model matrices}
\usage{
checkDesigns(designs)
}
\arguments{
\item{designs}{an alleged list of model matrices}
}
\value{
the list of model matrices, assuming it passes
}
\description{
verify that a list of matrices is in fact a named list of model matrices
}
\details{
this function will squawk and stop if the list is no good
}
