% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{get_pbmc3k_data}
\alias{get_pbmc3k_data}
\title{Load the pbmc3k dataset}
\usage{
get_pbmc3k_data()
}
\value{
Seurat object with \code{$cell_type} info in the \code{meta.data} slot.
}
\description{
This dataset is adapted directly from the Satija lab "pbmc3k" dataset used in their popular tutorial on guided clustering. It is provided in this package for convenience since "SeuratData" is not available on CRAN.

For more information, please see their documentation.
}
\details{
2,700 peripheral blood mononuclear cells (PBMC) from 10x genomics taken from the "SeuratData" package
}
