% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBestRank.R
\name{GetBestRank}
\alias{GetBestRank}
\title{determine the appropriate rank for an AutoNMF decomposition}
\usage{
GetBestRank(df, tol.overfit = 1e-04, ...)
}
\arguments{
\item{df}{a data.frame of output from crossvalidation: rep, rank, error}

\item{tol.overfit}{tolerance for increase in test set reconstruction error relative to minimum observed value during fitting}

\item{...}{not implemented}
}
\value{
the lowest rank that minimizes the reconstruction error
}
\description{
determine the appropriate rank for an AutoNMF decomposition
}
