% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesigns.R
\name{getDesigns}
\alias{getDesigns}
\title{Refactored out from AnnotateNMF to ease argument handling}
\usage{
getDesigns(columns = NULL, meta.data = NULL, designs = NULL, max.levels = 200)
}
\arguments{
\item{columns}{factor columns of meta.data, optional if !is.null(designs)}

\item{meta.data}{a data.frame of annotations, optional if !is.null(designs)}

\item{designs}{named list of design matrices (supersedes meta.data/columns)}

\item{max.levels}{maximum number of levels permitted for a factor to be kept}
}
\value{
a named list of design matrices, if one was not provided
}
\description{
Refactored out from AnnotateNMF to ease argument handling
}
