% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelMatrix.R
\name{getModelMatrix}
\alias{getModelMatrix}
\title{automatically generate a means model (one-vs-all group associations)}
\usage{
getModelMatrix(field, meta.data, sparse = FALSE)
}
\arguments{
\item{field}{the name of a column in the data.frame}

\item{meta.data}{a data.frame with one or more factor columns}

\item{sparse}{fit a sparse model.matrix? (FALSE)}
}
\value{
a model.matrix or sparse.model.matrix (if sparse==TRUE)
}
\description{
A little-known trick in limma is to fit ~ 0 + group for a means model.
This function automates that for a data.frame and a factor column of it.
}
\examples{

if (!exists("pbmc3k") | ! "nmf" \%in\% Reductions(pbmc3k)) { 
  get_pbmc3k_data() \%>\% NormalizeData \%>\% RunNMF \%>\% AnnotateNMF -> pbmc3k
} 


}
