% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelResults.R
\name{getModelResults}
\alias{getModelResults}
\title{extract data.frame of lods and pvalues for differential factor representation}
\usage{
getModelResults(fit, noneg = TRUE, noint = TRUE, tail = "pos")
}
\arguments{
\item{fit}{an lmFit result from limma, shrunken with eBayes()}

\item{noneg}{drop results with negative lods scores? (TRUE)}

\item{noint}{drop any results for '(Intercept)'? (TRUE)}

\item{tail}{add direction for one or two-tailed testing (pos,neg,std), default = pos}
}
\value{
a data.frame with columns 'factor', 'group', 'fc', and 'p'
}
\description{
log-odds of non-null differences for a response by a factor are in fit$lods
(which will usually be a matrix), and one-sided p-values for the moderated t
test are computed from fit$t and fit$df.total using pt(t, df, lower=FALSE),
then adjusted using the step-up procedure of Benjamini & Hochberg.
}
\details{
If an (Intercept) term is found, it will be dropped, and if
             negative LODS scores are encountered, they will be dropped,
             unless `noneg` and/or `noint` are FALSE.
}
