% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataHeatmap.R, R/MetadataSummary.R,
%   R/plot.nmf_metadata_summary.R
\name{MetadataHeatmap}
\alias{MetadataHeatmap}
\alias{MetadataSummary}
\alias{plot.nmf_metadata_summary}
\title{Summarize contribution of sample groups to NMF factors}
\usage{
MetadataHeatmap(x)

MetadataSummary(h, factor_data, reorder = TRUE)

\method{plot}{nmf_metadata_summary}(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{h}{matrix giving factors as rows and samples as columns}

\item{factor_data}{a factor of the same length as the number of columns in \code{h}}

\item{reorder}{sort results by proportion in each group (uses \code{hclust} if >2 groups)}

\item{...}{not implemented}
}
\value{
\code{data.frame} of mean weights for each sample group within each factor of class \code{nmf_metadata_summary}. Use the \code{plot} method to visualize.
}
\description{
Calculate the mean weight of samples in discrete and unique groups to each factor
}
