% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normit.R
\name{normit}
\alias{normit}
\title{given data in rankable order, squash it into 0-1 and project onto a Gaussian.}
\usage{
normit(x)
}
\arguments{
\item{x}{data that can be ranked somehow (usually numeric)}
}
\value{
the same data projected onto (-Inf, +Inf)
}
\description{
given data in rankable order, squash it into 0-1 and project onto a Gaussian.
}
\seealso{
normitAll for matrices or other rectangular shaped things
}
