% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc3k.R
\docType{data}
\name{pbmc3k}
\alias{pbmc3k}
\title{Compressed form of pbmc3k dataset}
\format{
compressed version of the \code{dgCMatrix}, use \code{\link{get_pbmc3k_data}} to use this dataset.
}
\usage{
data(pbmc3k)
}
\description{
See \code{\link{get_pbmc3k_data}}
}
\keyword{datasets}
