% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationDotPlot.R
\name{plotAnnotation}
\alias{plotAnnotation}
\title{Add two numbers together}
\usage{
plotAnnotation(dataset, datafield, transparency = NULL)
}
\arguments{
\item{x, y}{A pair of numbers.}
}
\description{
Add two numbers together
}
