% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{PreprocessData.Seurat}
\alias{PreprocessData.Seurat}
\alias{PreprocessData.Assay}
\alias{PreprocessData.dgCMatrix}
\alias{PreprocessData}
\title{Normalize count data}
\usage{
\method{PreprocessData}{Seurat}(object, scale.factor = 10000, assay = NULL, ...)

\method{PreprocessData}{Assay}(object, scale.factor = 10000, ...)

\method{PreprocessData}{dgCMatrix}(object, scale.factor = 10000, ...)

PreprocessData(object, scale.factor, ...)
}
\arguments{
\item{object}{Seurat object}

\item{scale.factor}{value by which to multiply all columns after unit normalization and before \code{log1p} transformation}

\item{assay}{assay in which the counts matrix resides}

\item{...}{not implemented}
}
\description{
Standard log-normalization equivalent to \code{Seurat::LogNormalize}, but faster and more memory-efficient
}
