% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_model.R
\name{project_model}
\alias{project_model}
\title{Project a factor model}
\usage{
project_model(A, w, L1 = 0.01, L2 = 0, threads = 0)
}
\arguments{
\item{A}{sparse matrix (ideally variance-stabilized) of data for genes x cells (rows x columns)}

\item{w}{matrix giving the factor model, of dimensions \code{nrow(A) x k}}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}
}
\value{
list of \code{h} and \code{d}, where \code{d} gives the relative contribution of each factor in \code{h} to the model
}
\description{
Project a dataset onto a factor model for transfer learning
}
