% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPlot.R
\name{RankPlot}
\alias{RankPlot}
\alias{RankPlot.Seurat}
\title{Plot NMF cross-validation results}
\usage{
RankPlot(object, reduction = "nmf", ...)

\method{RankPlot}{Seurat}(object, reduction = "nmf", detail.level = 1, ...)
}
\arguments{
\item{object}{a Seurat object or a \code{data.frame} that is the result of \code{RunNMF}}

\item{reduction}{the NMF reduction slot name (result of \code{RunNMF} where \code{k} was an array)}

\item{...}{not implemented}

\item{detail.level}{of detail to plot, \code{1} for test set reconstruction error at convergence of each factorization, \code{2} for test set reconstruction error at each fitting iteration of each factorization}
}
\value{
A ggplot2 object
}
\description{
Given a NMF reduction at multiple ranks, plot rank vs. test set reconstruction error to determine the optimal rank.

S3 method for Seurat that runs the \code{singlet::RunNMF} function.
}
