% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_rowwise.R
\name{RasterizeRowwise}
\alias{RasterizeRowwise}
\title{Bin together values from every block of \code{n} rows and calculate mean value, with a sparse \code{dgCMatrix} as input and a dense \code{matrix} as output. This technique is useful in some genomics applications.}
\usage{
RasterizeRowwise(A, n = 10, threads = 0)
}
\arguments{
\item{A}{matrix to be rasterized}

\item{n}{row-wise binning size}

\item{threads}{number of threads to use (0 to let OpenMP decide how many are available and use them all)}
}
\description{
Bin together values from every block of \code{n} rows and calculate mean value, with a sparse \code{dgCMatrix} as input and a dense \code{matrix} as output. This technique is useful in some genomics applications.
}
