% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescaleNMF.R
\name{RescaleNMF}
\alias{RescaleNMF}
\title{Rescaling of cell and gene features}
\usage{
RescaleNMF(object, ...)
}
\arguments{
\item{object}{is a seurat object}

\item{...}{other unknown parameters can be added to then function.}

\item{features}{are the list of cell or genes of interest.}

\item{reduction}{use to specifiy the reduction .}

\item{lambda}{is the threshold parameter between 0 and 1 which helps improve the relationship of v and either H or W.}

\item{reduction.name}{names in which the new rescaled seurate object is stored.}
}
\description{
Rescaling of cell and gene features
}
\examples{
RescaleNMF(pbmc3k,c("AGRN","dvl1"),0.4)
}
