% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescaleSpatial.R
\name{RescaleSpatial.Seurat}
\alias{RescaleSpatial.Seurat}
\alias{RescaleSpatial}
\title{Rescale spatial coordinates}
\usage{
\method{RescaleSpatial}{Seurat}(object, reduction = "spatial")
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{the name of the spatial reduction to use}
}
\value{
Seurat object with rescaled spatial coordinates
}
\description{
Convert coordinates in the "spatial" reduction to natural numbers rather than values between 0 and 1. This allows for intuitive graph construction based on the radius surrounding any given cell (i.e. a radius of one corresponds to all cells next to the cell of interest)
}
