% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunLNMF.R
\name{run_linked_nmf}
\alias{run_linked_nmf}
\title{Run Linked Non-negative Matrix Factorization}
\usage{
run_linked_nmf(
  A,
  w,
  link_h = NULL,
  link_w = NULL,
  tol = 1e-04,
  maxit = 100,
  verbose = TRUE,
  L1 = 0.01,
  L2 = 0,
  threads = 0
)
}
\arguments{
\item{A}{sparse matrix giving normalized counts for genes x cells (rows x columns), or a list of sparse matrices with equal number of rows and identical rownames}

\item{w}{initial matrix for 'w', usually taken from the result of \code{run_nmf}, of dimensions \code{nrow(A) x rank}.}

\item{link_h}{matrix giving the linkage weight (usually in the range \code{(0, 1)}) of dimensions \code{rank x ncol(A)}.}

\item{link_w}{matrix giving the linkage weight of dimensions \code{nrow(A) x rank}.}

\item{tol}{tolerance of the fit (1e-5 for publication quality, 1e-4 for cross-validation)}

\item{maxit}{maximum number of iterations}

\item{verbose}{verbosity level}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}
}
\description{
Run LNMF, initialized from any NMF model, where factors may be "linked" to certain samples.
}
