% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescaleSpatial.R, R/RunGCNMF.R
\name{RunGCNMF}
\alias{RunGCNMF}
\alias{RescaleSpatial}
\alias{RunGCNMF.Seurat}
\title{Run Graph-Convolutional Non-negative Matrix Factorization}
\usage{
RescaleSpatial(object, ...)

\method{RunGCNMF}{Seurat}(
  object,
  graph,
  k,
  split.by = NULL,
  assay = NULL,
  tol = 1e-05,
  L1 = 0.01,
  L2 = 0,
  verbose = 2,
  reduction.name = "gcnmf",
  reduction.key = "GCNMF_",
  maxit = 100,
  threads = 0,
  features = NULL,
  ...
)

RunGCNMF(object, ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{...}{not implemented}

\item{graph}{A graph to use, either directed or undirected}

\item{k}{rank of the factorization (no automatic rank determination for GCNMF. Use \code{\link{RunNMF}}). Alternatively, specify an initial \code{w} matrix of dimensions \code{m x k}, where \code{m} is the number of rows in the matrix to be factorized.}

\item{split.by}{column name in \code{colData} giving a \code{factor} with multiple levels for splitting. Data will be weighted such that each level in the factor contributes equally to the NMF model.}

\item{assay}{Assay to use, defaults to logcounts}

\item{tol}{tolerance of the fit (correlation distance of the model across consecutive iterations). Cross-validation fits are 10x coarser than this tolerance.}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{L2}{L2/Ridge-like penalty to increase angles between factors}

\item{verbose}{print updates to console}

\item{reduction.name}{Name to store resulting DimReduc object as ("NMF")}

\item{reduction.key}{Key for resulting DimReduc ("NMF")}

\item{maxit}{maximum number of fitting iterations}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}

\item{features}{unused for this method}
}
\value{
Returns a Seurat object with the GCNMF model stored in the reductions slot
}
\description{
Run NMF with weighted convolution determined by edges in a graph of dimensions \code{n x n}, where \code{n} is the number of columns in the matrix.
}
\details{
Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunNMF}}
}
