% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGSEA.R
\name{RunGSEA}
\alias{RunGSEA}
\alias{RunGSEA.Seurat}
\alias{RunGSEA.SingleCellExperiment}
\alias{RunGSEA.list}
\alias{RunGSEA.nmf}
\alias{RunGSEA.matrix}
\title{Run Gene Set Enrichment Analysis on a Reduction}
\usage{
RunGSEA(object, ...)

\method{RunGSEA}{Seurat}(
  object,
  ID = "gene_symbol",
  reduction = "nmf",
  species = "Homo sapiens",
  category = "C5",
  min.size = 10,
  max.size = 500,
  dims = NULL,
  verbose = TRUE,
  padj.sig = 0.01,
  add.noise = FALSE,
  scoreType = "pos",
  gsea.name = "gsea",
  remove.zeros.per.factor = FALSE,
  ...
)

\method{RunGSEA}{SingleCellExperiment}(
  object,
  ID = "gene_symbol",
  reduction = "nmf",
  species = "Homo sapiens",
  category = "C5",
  min.size = 10,
  max.size = 500,
  dims = NULL,
  verbose = TRUE,
  padj.sig = 0.01,
  add.noise = FALSE,
  scoreType = "pos",
  gsea.name = "gsea",
  remove.zeros.per.factor = FALSE,
  ...
)

\method{RunGSEA}{list}(
  object,
  ID = "gene_symbol",
  reduction = "nmf",
  species = "Homo sapiens",
  category = "C5",
  min.size = 10,
  max.size = 500,
  dims = NULL,
  verbose = TRUE,
  padj.sig = 0.01,
  add.noise = FALSE,
  scoreType = "pos",
  gsea.name = "gsea",
  remove.zeros.per.factor = FALSE,
  ...
)

\method{RunGSEA}{nmf}(
  object,
  ID = "gene_symbol",
  reduction = "nmf",
  species = "Homo sapiens",
  category = "C5",
  min.size = 10,
  max.size = 500,
  dims = NULL,
  verbose = TRUE,
  padj.sig = 0.01,
  add.noise = FALSE,
  scoreType = "pos",
  gsea.name = "gsea",
  remove.zeros.per.factor = FALSE,
  ...
)

\method{RunGSEA}{matrix}(
  w,
  ID = "gene_symbol",
  reduction = "nmf",
  species = "Homo sapiens",
  category = "C5",
  min.size = 10,
  max.size = 500,
  dims = NULL,
  verbose = TRUE,
  padj.sig = 0.01,
  add.noise = FALSE,
  scoreType = "pos",
  remove.zeros.per.factor = FALSE,
  ...
)
}
\arguments{
\item{object}{a Seurat or RcppML::nmf object}

\item{...}{additional params to pass to msigdbr}

\item{ID}{gene ID format (i.e. "gene_symbol", "ensembl_gene")}

\item{reduction}{dimensional reduction to use (if Seurat)}

\item{species}{species for which to load gene sets}

\item{category}{msigdbr gene set category (i.e. "H", "C5", etc.)}

\item{min.size}{minimum number of terms in a gene set}

\item{max.size}{maximum number of terms in a gene set}

\item{dims}{factors in the reduction to use, default \code{NULL} for all factors}

\item{verbose}{print progress to console}

\item{padj.sig}{significance cutoff for BH-adjusted p-values (default 0.01)}

\item{add.noise}{add random noise to ranks to prevent hanging, known bug in fgsea(https://github.com/alserglab/fgsea/issues/151#issuecomment-2088857387)}

\item{gsea.name}{gsea name, gsea by default}

\item{remove.zeros.per.factor}{option to remove genes that have weight 0 within factor (not recommended unless you can prove this is unbiased filtering)}
}
\value{
a Seurat or nmf object, with GSEA information in the misc slot. BH-adj p-values are on a -log10 scale.
}
\description{
Run GSEA to identify gene sets that are enriched within NMF factors.
}
