% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{RunGSEA.Seurat}
\alias{RunGSEA.Seurat}
\title{Run Gene Set Enrichment Analysis}
\usage{
RunGSEA.Seurat(object, reduction = "nmf")
}
\description{
Perform GSEA on feature loadings in a reduction contained within a Seurat object (i.e. NMF factors) using MSIGDBR pathways or manually provided pathways.
}
