% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{RunLNMF.Seurat}
\alias{RunLNMF.Seurat}
\alias{RunLNMF}
\alias{MetadataPlot}
\alias{MetadataPlot.Seurat}
\alias{GetSharedFactors}
\alias{GetUniqueFactors}
\title{Run Linked NMF on a Seurat object}
\usage{
\method{RunLNMF}{Seurat}(
  object,
  split.by,
  reduction.use = "nmf",
  reduction.name = "lnmf",
  reduction.key = "LNMF_",
  verbose = TRUE,
  link.cutoff = 0.5,
  tol = 1e-04,
  maxit = 100,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  precision = "double",
  link.balance.tol = 0,
  balance.maxit = 100,
  link.balance.rate = 0.1,
  ...
)

RunLNMF(object, ...)

MetadataPlot(object, ...)

\method{MetadataPlot}{Seurat}(object, split.by, reduction = "lnmf", ...)

GetSharedFactors(object, split.by, reduction = "lnmf")

GetUniqueFactors(object, split.by, reduction = "lnmf")
}
\arguments{
\item{object}{A Seurat object}

\item{split.by}{attribute in \code{@metadata} for splitting. Data will be weighted such that each group contributes equally to the LNMF model.}

\item{reduction.use}{NMF reduction to use for initializing the linked factorization.}

\item{reduction.name}{name to store resulting DimReduc object as}

\item{reduction.key}{key for resulting DimReduc}

\item{verbose}{print fitting progress to console}

\item{link.cutoff}{if the relative contribution of samples in any given group to a factor falls below \code{link.cutoff}, unlink it from the factor. \code{link.cutoff = 1} means a factor must contribute exactly equally before being unlinked.}

\item{tol}{tolerance of the fit (correlation distance of the model across consecutive iterations)}

\item{maxit}{maximum number of fitting iterations}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{L2}{L2/Ridge penalty to increase angles between factors and sparsity of the model}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}

\item{precision}{\code{"double"} or \code{"float"} for numerical precision. \code{"float"} may be faster, but numerical instability may result in more iterations to achieve desired tolerances.}

\item{link.balance.tol}{relative change in factor representation within sample groups between balancing iterations at which to call convergence.}

\item{balance.maxit}{maximum number of iterations for the balancing step}

\item{link.balance.rate}{proportion of difference between current factor weight and equal representation of factors in a sample group to target in a single iteration (default 0.1).}

\item{...}{not implemented}

\item{reduction}{reduction to use for metadata analysis}

\item{link.balance}{after initial linking step, weight all shared factors such that each dataset is represented as equally as possible in each factor.}
}
\value{
Returns a Seurat object with the NMF model stored in the reductions slot
}
\description{
Run a Linked Non-negative Matrix Factorization to separate shared and unique signals for integration or signature extraction.

S3 method for Seurat that runs the \code{singlet::RunLNMF} function.
}
\details{
Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunNMF}}, \code{\link{RankPlot}}, \code{\link{MetadataSummary}}
}
