% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlet.R
\docType{package}
\name{singlet}
\alias{singlet}
\alias{singlet-package}
\title{Singlet}
\description{
Fast single-cell analysis with non-negative dimensional reductions
}
\details{
There are reasons to not use PCA.
\itemize{
\item PCA fits to missing signal,
\item considers only highly variable features,
\item is almost useless without further graph-based analysis,
\item requires centering and scaling of your data,
\item and is robust only within experiments.
}

Instead, you should use Non-negative Matrix Factorization (NMF).
\itemize{
\item NMF imputes missing signal,
\item learns models using all features,
\item does everything PCA does and provides useful information itself,
\item requires only variance stabilization,
\item and is robust across experiments.
}

Singlet is all about extremely fast NMF for single-cell dimensional reduction and integration.

See the vignettes to get started.
}
\author{
Zach DeBruine
}
