% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViziumPlot.R
\name{ViziumPlot}
\alias{ViziumPlot}
\title{ViziumPlot}
\usage{
ViziumPlot(
  x_coords,
  y_coords,
  fill,
  upper_right_edge,
  right_edge,
  lower_right_edge,
  lower_left_edge,
  left_edge,
  upper_left_edge,
  edge_width
)
}
\description{
Plots 10x Vizium data using hexagons. Edge weights can be controlled
}
