% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_IVCSC}
\alias{write_IVCSC}
\title{Write an IVCSC matrix}
\usage{
write_IVCSC(L, verbose = TRUE)
}
\arguments{
\item{L}{input dgCMatrix list}

\item{verbose}{print outputs}
}
\description{
Write an IVCSC matrix
}
