% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateNMF.R
\name{AnnotateNMF}
\alias{AnnotateNMF}
\alias{AnnotateNMF.DimReduc}
\alias{AnnotateNMF.Seurat}
\alias{AnnotateNMF.nmf}
\title{annotate an NMF model}
\usage{
AnnotateNMF(object, ...)

\method{AnnotateNMF}{DimReduc}(object, meta.data, ...)

\method{AnnotateNMF}{Seurat}(object, fields = NULL, reduction = "nmf", ...)

\method{AnnotateNMF}{nmf}(object, meta.data, ...)
}
\arguments{
\item{object}{an object suitable for annotation (Seurat, DimReduc, or nmf)}

\item{meta.data}{a data.frame}

\item{fields}{fields (columns) of meta.data to annotate against}

\item{reduction}{the reductions slot in the Seurat object containing the model to annotate}
}
\description{
This function clearly expects meta.data to be a data.frame with cells as rows
}
\details{
Maps factor information from the \code{meta.data} slot of a Seurat object or a specified factor to each NMF factor, and computes summary statistics

Maps factor information in an RcppML::nmf object against meta.data
}
\examples{
if (!exists("pbmc3k")) get_pbmc3k_data() \%>\% NormalizeData -> pbmc3k
if (!"nmf" \%in\% Reductions(pbmc3k)) pbmc3k \%>\% RunNMF() -> pbmc3k
AnnotateNMF(pbmc3k)

}
