% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateNMF.R
\name{AnnotateNMF}
\alias{AnnotateNMF}
\alias{AnnotateNMF.DimReduc}
\alias{AnnotateNMF.Seurat}
\alias{AnnotateNMF.nmf}
\title{annotate an NMF model}
\usage{
AnnotateNMF(object, ...)

\method{AnnotateNMF}{DimReduc}(
  object,
  meta.data = NULL,
  columns = NULL,
  designs = NULL,
  center = TRUE,
  scale = FALSE,
  max.levels = 200,
  tail = "pos",
  annotation.name = "annotations",
  ...
)

\method{AnnotateNMF}{Seurat}(
  object,
  columns = NULL,
  reduction = "nmf",
  tail = "pos",
  annotation.name = "annotations",
  ...
)

\method{AnnotateNMF}{nmf}(
  object,
  meta.data,
  columns = NULL,
  designs = NULL,
  center = TRUE,
  scale = FALSE,
  max.levels = 200,
  tail = "pos",
  annotation.name = "annotations",
  ...
)
}
\arguments{
\item{object}{an object suitable for annotation (Seurat, DimReduc, or nmf)}

\item{...}{not implemented}

\item{meta.data}{a data.frame, if one is not already part of the object}

\item{columns}{factor columns of meta.data (see below) to annotate against}

\item{designs}{named list of design matrices (supersedes meta.data/columns)}

\item{center}{center the factor matrix for testing? (TRUE)}

\item{scale}{scale the factor matrix for testing? (FALSE)}

\item{max.levels}{maximum number of levels a factor may have in order to be included in analysis}

\item{tail}{add direction for one or two-tailed testing (pos,neg,std), default = pos}

\item{annotation.name}{name of factor annotation, default = "annotations"}

\item{reduction}{the reductions slot in the Seurat object containing the model to annotate}
}
\description{
annotate an NMF model

Annotate NMF model with cell or sample metadata

Annotate NMF model with cell metadata
}
\details{
Maps factor information in an RcppML::nmf object against meta.data
}
\examples{
\dontrun{
get_pbmc3k_data() \%>\%
  NormalizeData() \%>\%
  RunNMF() -> pbmc3k
AnnotateNMF(pbmc3k)
}
}
