% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateNMF.R
\name{AnnotateNMF}
\alias{AnnotateNMF}
\alias{AnnotateNMF.DimReduc}
\alias{AnnotateNMF.Seurat}
\alias{AnnotateNMF.nmf}
\title{annotate an NMF model}
\usage{
AnnotateNMF(object, ...)

\method{AnnotateNMF}{DimReduc}(object, meta.data, shrink = TRUE, ...)

\method{AnnotateNMF}{Seurat}(object, fields = NULL, reduction = "nmf", ...)

\method{AnnotateNMF}{nmf}(object, meta.data, shrink = TRUE, ...)
}
\arguments{
\item{object}{an object suitable for annotation (Seurat, DimReduc, or nmf)}

\item{meta.data}{a data.frame}

\item{shrink}{apply eBayes to the fits? (TRUE)}

\item{fields}{fields (columns) of meta.data to annotate against}

\item{reduction}{the reductions slot in the Seurat object containing the model to annotate}
}
\description{
Note: some of the documentation (e.g. meta.data is a named list) is wrong.
This function clearly expects meta.data to be a data.frame with cells as rows

Note: some of the documentation (e.g. meta.data is a named list) is wrong.
}
\details{
Maps factor information from the \code{meta.data} slot of a Seurat object or a specified factor to each NMF factor, and computes summary statistics

Maps factor information in an RcppML::nmf object against meta.data
}
\examples{
if (!exists("pbmc3k") | 
    !"nmf" \%in\% Reductions(pbmc3k) | 
    !"annotations" \%in\% names(object@misc)) { 
  get_pbmc3k_data() \%>\% NormalizeData \%>\% RunNMF \%>\% AnnotateNMF -> pbmc3k
}
show(pbmc3k)

}
