% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateNMF.R, R/AnnotationPlot.R
\name{AnnotateNMF.DimReduc}
\alias{AnnotateNMF.DimReduc}
\alias{AnnotateNMF}
\alias{AnnotateNMF.Seurat}
\alias{AnnotationPlot}
\alias{AnnotationPlot.Seurat}
\alias{AnnotationPlot.DimReduc}
\title{Annotate NMF model with cell metadata}
\usage{
\method{AnnotateNMF}{DimReduc}(object, meta.data, shrink = TRUE, ...)

\method{AnnotateNMF}{Seurat}(object, fields = NULL, reduction = "nmf", ...)

AnnotationPlot(object, ...)

\method{AnnotationPlot}{Seurat}(object, plot.field = NULL, reduction = "nmf", ...)

\method{AnnotationPlot}{DimReduc}(object, plot.field = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{meta.data}{if a DimReduc object is specified, provide a named list containing factor(s) for which to calculate summary statistics against the NMF model}

\item{shrink}{apply eBayes to the fits? (TRUE)}

\item{...}{not implemented}

\item{fields}{if a Seurat object is specified, fields in the \code{meta.data} slot may be specified for which to compute summary statistics. If \code{fields = NULL}, all fields that are a \code{factor} with more than one level are considered.}

\item{reduction}{the reductions slot in the Seurat object containing the model to annotate}

\item{plot.field}{name of field in \code{meta.data} to plot, if \code{NULL}, plots the first field in the annotation results}
}
\value{
ggplot2 object
}
\description{
Note: some of the documentation (e.g. meta.data is a named list) is wrong.
This function clearly expects meta.data to be a data.frame with cells as rows

After running \code{AnnotateNMF}, this function returns 
a dot plot of the results
}
\details{
Maps factor information from the \code{meta.data} slot of a Seurat object or a specified factor to each NMF factor, and computes summary statistics
}
\examples{
if (!exists("pbmc3k") | ! "nmf" \%in\% Reductions(pbmc3k)) { 
  get_pbmc3k_data() \%>\% NormalizeData() \%>\% RunNMF() -> pbmc3k
}
(pbmc3k <- AnnotateNMF(pbmc3k))

if (!exists("pbmc3k") | ! "nmf" \%in\% Reductions(pbmc3k) | ) { 
  get_pbmc3k_data() \%>\% NormalizeData() \%>\% RunNMF() -> pbmc3k
}
pbmc3k \%>\% AnnotateNMF() \%>\% AnnotationPlot()

}
