% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPlot.R
\name{AnnotationPlot}
\alias{AnnotationPlot}
\alias{AnnotationPlot.Seurat}
\alias{AnnotationPlot.DimReduc}
\alias{AnnotationPlot.nmf}
\alias{AnnotationPlot.data.frame}
\title{Plot annotations from an NMF model or other compatible objects.}
\usage{
AnnotationPlot(object, ...)

\method{AnnotationPlot}{Seurat}(object, plot.field = NULL, reduction = "nmf", ...)

\method{AnnotationPlot}{DimReduc}(object, plot.field = NULL, ...)

\method{AnnotationPlot}{nmf}(object, plot.field = NULL, ...)

\method{AnnotationPlot}{data.frame}(object, plot.field, ...)
}
\arguments{
\item{object}{a compatible object (Seurat, DimReduc, nmf, data.frame)}

\item{...}{additional arguments passed to called functions}

\item{plot.field}{metadata grouping to plot}

\item{reduction}{the reduction to plot (default is 'nmf')}
}
\value{
a ggplot2 object
}
\description{
After running \code{AnnotateNMF}, this function returns 
a dot plot of the results

After running \code{AnnotateNMF}, this function returns 
a dot plot of the results.  Right now the code is the same as for DimReduc.
}
\examples{

if (!exists("pbmc3k")) get_pbmc3k_data() \%>\% NormalizeData -> pbmc3k
if (!"nmf" \%in\% Reductions(pbmc3k)) pbmc3k \%>\% RunNMF() -> pbmc3k
AnnotateNMF(pbmc3k) \%>\% AnnotationPlot("cell_type")

dat <- pbmc3k@reductions$nmf@misc$annotations$cell_type
AnnotationPlot(dat, "cell_type")

}
