% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{ard_nmf}
\alias{ard_nmf}
\title{Automatic Rank Determination NMF}
\usage{
ard_nmf(
  A,
  k_init = 10,
  n_replicates = 3,
  tol = 1e-05,
  maxit = 100,
  verbose = 1,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  test_density = 0.05,
  learning_rate = 0.8,
  tol_overfit = 1e-04,
  trace_test_mse = 5,
  detail_level = 1
)
}
\arguments{
\item{A}{sparse matrix (ideally variance-stabilized) of data for genes x cells (rows x columns)}

\item{n_replicates}{number of random test sets}

\item{tol}{tolerance of the fit (1e-5 for publication quality, 1e-3 for cross-validation)}

\item{maxit}{maximum number of iterations}

\item{verbose}{no output (0/FALSE), rank-level output (1/TRUE) and step size info (2) and individual model fitting updates (3)}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}

\item{test_density}{fraction of values to include in the test set}

\item{learning_rate}{exponent on step size for automatic rank determination}

\item{tol_overfit}{stopping criterion, maximum increase in test set reconstruction error at any iteration compared to test set reconstruction error at \code{trace_test_mse}}

\item{trace_test_mse}{first iteration at which to calculate test set reconstruction error, and the error to compare all later iterations to when determining whether overfitting has occurred.}

\item{detail_level}{\code{1} for returning best test set reconstruction error for each factorization, \code{2} to return test set reconstruction error at all \code{trace_test_mse} iterations in each factorization}
}
\description{
ARD NMF quickly finds the optimal rank for an NMF model using an exponentially variable learning rate and basic coordinate descent.
}
