% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{cross_validate_nmf}
\alias{cross_validate_nmf}
\alias{plot.cross_validate_nmf_data}
\title{Determine best rank for NMF using cross-validation}
\usage{
cross_validate_nmf(
  A,
  ranks,
  n_replicates = 3,
  tol = 1e-04,
  maxit = 100,
  verbose = 1,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  test_density = 0.05,
  tol_overfit = 1e-04,
  trace_test_mse = 5,
  detail_level = 1
)

\method{plot}{cross_validate_nmf_data}(x, detail = 2, ...)
}
\arguments{
\item{A}{sparse matrix (ideally variance-stabilized) of data for genes x cells (rows x columns)}

\item{ranks}{a vector of ranks at which to fit a model and compute test set reconstruction error}

\item{n_replicates}{number of random test sets}

\item{tol}{tolerance of the fit (1e-5 for publication quality, 1e-3 for cross-validation)}

\item{maxit}{maximum number of iterations}

\item{verbose}{verbosity level}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}

\item{test_density}{fraction of values to include in the test set}

\item{tol_overfit}{stopping criterion, maximum increase in test set reconstruction error at any iteration compared to test set reconstruction error at \code{trace_test_mse}}

\item{trace_test_mse}{first iteration at which to calculate test set reconstruction error, and the error to compare all later iterations to when determining whether overfitting has occurred.}

\item{detail_level}{\code{1} for returning best test set reconstruction error for each factorization, \code{2} to return test set reconstruction error at all \code{trace_test_mse} iterations in each factorization}

\item{x}{the result of \code{cross_validate_nmf}}

\item{detail}{level of detail to plot}

\item{...}{additional arguments (not implemented)}
}
\value{
a \code{data.frame} of test set reconstruction error vs. rank of class \code{nmf_cross_validate_data}. Use \code{plot} method to visualize or \code{min} to compute optimal rank.
}
\description{
Find the rank that minimizes the mean squared error of test set reconstruction using cross-validation.
}
