% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBestRank.R
\name{GetBestRank}
\alias{GetBestRank}
\title{determine the appropriate rank for an ARD ("RADAR") NMF decomposition}
\usage{
GetBestRank(df)
}
\arguments{
\item{df}{a data.frame of output from crossvalidation: rep, rank, error}
}
\value{
the lowest rank that minimizes the reconstruction error
}
\description{
determine the appropriate rank for an ARD ("RADAR") NMF decomposition
}
