% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesigns.R
\name{getDesigns}
\alias{getDesigns}
\title{refactored out from AnnotateNMF to ease argument handling}
\usage{
getDesigns(columns = NULL, meta.data = NULL, designs = NULL)
}
\arguments{
\item{columns}{factor columns of meta.data, optional if !is.null(designs)}

\item{meta.data}{a data.frame of annotations, optional if !is.null(designs)}

\item{designs}{named list of design matrices (supersedes meta.data/columns)}

\item{...}{not implemented}
}
\value{
a named list of design matrices, if one was not provided
}
\description{
refactored out from AnnotateNMF to ease argument handling
}
