% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelFit.R
\name{getModelFit}
\alias{getModelFit}
\title{get linear all-pairs comparisons fits for a design matrix and data matrix}
\usage{
getModelFit(design, object, shrink = TRUE)
}
\arguments{
\item{design}{a model.matrix (or a sparse.model.matrix, perhaps)}

\item{object}{a data.matrix or a Seurat DimReduc object}

\item{shrink}{run limma::eBayes() on the model? (TRUE)}
}
\description{
Continuing along with the theme of "stupid limma tricks", this function 
fits and (optionally, though by default) shrinks a linear model for a factor.
}
