% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelFit.R
\name{getModelFit}
\alias{getModelFit}
\title{get linear all-pairs comparisons fits for a design matrix and data matrix}
\usage{
getModelFit(design, object, center = TRUE, ...)
}
\arguments{
\item{design}{a model.matrix (or a sparse.model.matrix, perhaps)}

\item{object}{a data.matrix, Seurat DimReduc, or RcppML nmf object}

\item{center}{center the factor matrix for testing? (TRUE)}

\item{...}{not implemented}
}
\description{
Continuing along with the theme of "stupid limma tricks", this function 
fits and shrinks a means model for a factor. The proportion of factors
assumed to have a fold-change > 0 is 1%, and a robust fit is applied.
}
