% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelMatrix.R
\name{getModelMatrix}
\alias{getModelMatrix}
\title{automatically generate a means model (one-vs-all group associations)}
\usage{
getModelMatrix(field, meta.data = NULL, sparse = FALSE, ova = TRUE, ...)
}
\arguments{
\item{field}{the name of a column in the data.frame, or the column}

\item{meta.data}{a data.frame with one or more factor columns, or NULL}

\item{sparse}{fit a sparse model.matrix? (FALSE)}

\item{ova}{fit a One-Vs-All model matrix (no referent)? (TRUE)}

\item{...}{any additional params to pass to model.matrix}
}
\value{
a model.matrix or sparse.model.matrix (if sparse==TRUE)
}
\description{
A little-known trick in limma is to fit ~ 0 + group for a means model.
This function automates that for a data.frame and a factor column of it.
}
\details{
If a factor (and no meta.data) is supplied (usually by with(meta.data, ...)),
getModelMatrix will attempt to figure out the text to remove from the matrix
column names by using deparse() and match.call() on the arguments (voodoo!).
In order to fit one-vs-all comparisons, a means model is the default. If you
have a referent group (e.g. normal bone marrow vs. a bunch of leukemia cells)
or simply don't want a means model, set `ova` (one vs all) to FALSE.
}
