% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelResults.R
\name{getModelResults}
\alias{getModelResults}
\title{extract data.frame of lods and pvalues for differential factor representation}
\usage{
getModelResults(fit)
}
\arguments{
\item{fit}{an lmFit result from limma, shrunken with eBayes()}
}
\value{
a data.frame with columns 'factor', 'group', 'fc', and 'p'
}
\description{
log-odds of non-null differences for a response by a factor are in fit$lods
(which will usually be a matrix), and one-sided p-values for the moderated t
test are computed from fit$t and fit$df.total using pt(t, df, lower=FALSE),
then adjusted using the step-up procedure of Benjamini & Hochberg.
}
