% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelResults.R
\name{getModelResults}
\alias{getModelResults}
\title{extract data.frame of lods and pvalues for differential factor representation}
\usage{
getModelResults(fit)
}
\arguments{
\item{fit}{an lmFit result from limma, perhaps shrunken with eBayes()}
}
\value{
a data.frame with columns 'factor', 'group', 'fc', and 'p'
}
\description{
log-odds of non-null differences for a response by a factor are in fit$lods
(which will usually be a matrix) and p-values will be in fit$p.value (these
are raw, so we adjust them to yield a B-H fdr). The results are then merged.
}
