% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{GSEAHeatmap}
\alias{GSEAHeatmap}
\title{Plot GSEA results on a heatmap}
\usage{
GSEAHeatmap(object, reduction = "nmf", max.terms.per.factor = 3)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{a dimensional reduction for which GSEA analysis has been performed}

\item{max.terms.per.factor}{show this number of top terms for each factor}
}
\value{
ggplot2 object
}
\description{
Plot top GSEA terms for each NMF factor on a heatmap
}
