% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{MetadataSummary}
\alias{MetadataSummary}
\alias{plot.nmf_metadata_summary}
\alias{MetadataHeatmap}
\title{Summarize contribution of sample groups to NMF factors}
\usage{
MetadataSummary(h, factor_data, reorder = TRUE)

\method{plot}{nmf_metadata_summary}(x, ...)

MetadataHeatmap(x)
}
\arguments{
\item{h}{matrix giving factors as rows and samples as columns}

\item{factor_data}{a factor of the same length as the number of columns in \code{h}}

\item{reorder}{sort results by proportion in each group (uses \code{hclust} if >2 groups)}

\item{x}{result of \code{MetadataSummary}}

\item{...}{not implemented}
}
\value{
\code{data.frame} of mean weights for each sample group within each factor of class \code{nmf_metadata_summary}. Use the \code{plot} method to visualize.
}
\description{
Calculate the mean weight of samples in discrete and unique groups to each factor
}
