% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactorWeights.R
\name{plotFactorWeights}
\alias{plotFactorWeights}
\title{convenience function to plot one or more factors along a genome using igvR}
\usage{
plotFactorWeights(object, gr, factors = 1:3, plot = NULL)
}
\arguments{
\item{object}{an nmf object or something with a @w weights matrix}

\item{gr}{a GRanges object with coordinates for the features}

\item{factors}{which factors to plot weights for (default: 1, 2, 3)}

\item{plot}{use igvR to plot the factors? (TRUE, if igvR detected)}
}
\value{
the GRanges gr, but with factor weights added as mcols
}
\description{
convenience function to plot one or more factors along a genome using igvR
}
