% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectData.Seurat}
\alias{ProjectData.Seurat}
\alias{ProjectData}
\alias{ProjectData.SingleCellExperiment}
\title{Project data onto a factor model}
\usage{
\method{ProjectData}{Seurat}(
  object,
  w,
  split.by = NULL,
  assay = NULL,
  L1 = 0.01,
  L2 = 0,
  reduction.name = "nmf_projection",
  reduction.key = "NNLS_",
  threads = 0,
  reorder = FALSE,
  ...
)

\method{ProjectData}{SingleCellExperiment}(
  object,
  w,
  split.by = NULL,
  assay = "logcounts",
  L1 = 0.01,
  L2 = 0,
  reduction.name = "NMF",
  reduction.key = "NMF_",
  threads = 0,
  reorder = FALSE,
  ...
)

ProjectData(object, ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{w}{factor loadings with nrow(w) equal to nrow(object)}

\item{split.by}{column name in \code{colData} giving a \code{factor} with multiple levels for splitting. Data will be weighted such that each level in the factor contributes equally to the NMF model.}

\item{assay}{Assay to use, defaults to logcounts}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{L2}{L2/Ridge-like penalty to increase angles between factors}

\item{reduction.name}{Name to store resulting DimReduc object as ("NMF")}

\item{reduction.key}{Key for resulting DimReduc ("NMF")}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}

\item{reorder}{reorder the factors of the projection by d? (FALSE)}

\item{...}{not implemented}
}
\value{
Returns a Seurat object with the projection stored in the reductions slot

a SingleCellExperiment with projection stored in reducedDim(, "NNLS")
}
\description{
Non-negative Least Squares (NNLS) projection of assay data onto a factor model for transfer learning

Non-negative Least Squares (NNLS) projection of assay data onto a factor model for transfer learning
}
\details{
Use \code{set.seed()} to guarantee reproducibility!

Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunLNMF}}, \code{\link{MetadataSummary}}

\code{\link{RunLNMF}}, \code{\link{MetadataSummary}}
}
