% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectData.Seurat}
\alias{ProjectData.Seurat}
\alias{ProjectData}
\title{Project data onto a factor model}
\usage{
\method{ProjectData}{Seurat}(
  object,
  w,
  split.by = NULL,
  assay = NULL,
  L1 = 0.01,
  L2 = 0,
  reduction.name = "nmf_projection",
  reduction.key = "NNLS_",
  threads = 0,
  ...
)

ProjectData(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{w}{a factor model of the same number of rows as the assay to be projected in the current object, where columns correspond to factors}

\item{split.by}{column name in \code{@meta.data} giving a \code{factor} with multiple levels for splitting. Data will be weighted such that each level in the factor contributes equally to the NMF model.}

\item{assay}{Assay to use, defaults to the default assay of the first object}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{L2}{L2/Ridge-like penalty to increase angles between factors}

\item{reduction.name}{Name to store resulting DimReduc object as}

\item{reduction.key}{Key for resulting DimReduc}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}

\item{...}{not implemented}
}
\value{
Returns a Seurat object with the projection stored in the reductions slot
}
\description{
Non-negative Least Squares (NNLS) projection of assay data onto a factor model for transfer learning
}
\details{
Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunLNMF}}, \code{\link{MetadataSummary}}
}
