% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{RankPlot.Seurat}
\alias{RankPlot.Seurat}
\alias{RankPlot}
\title{Plot NMF cross-validation results}
\usage{
\method{RankPlot}{Seurat}(object, reduction = "nmf")

RankPlot(object, reduction = "nmf")
}
\arguments{
\item{object}{a Seurat object or a \code{data.frame} that is the result of \code{RunNMF}}

\item{reduction}{name of the NMF reduction in the Seurat object (result of \code{RunNMF}) for which multiple \code{ranks} were computed.}
}
\value{
A ggplot2 object
}
\description{
Given a NMF reduction at multiple ranks, plot rank vs. test set reconstruction error to determine the optimal rank.

S3 method for Seurat that runs the \code{singlet::RunNMF} function.
}
