% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{run_nmf}
\alias{run_nmf}
\title{Run Non-negative Matrix Factorization}
\usage{
run_nmf(
  A,
  rank,
  tol = 1e-04,
  maxit = 100,
  verbose = TRUE,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  precision = "float"
)
}
\arguments{
\item{A}{sparse matrix (ideally variance-stabilized) of data for genes x cells (rows x columns)}

\item{rank}{factorization rank}

\item{tol}{tolerance of the fit (1e-5 for publication quality, 1e-3 for cross-validation)}

\item{maxit}{maximum number of iterations}

\item{verbose}{verbosity level}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors and sparsity of the model}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}

\item{precision}{\code{"float"} or \code{"double"}. High tolerance models may not be achievable with \code{"float"} precision.}
}
\description{
Run NMF on a sparse matrix
}
