% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nmf.R
\name{run_nmf}
\alias{run_nmf}
\title{Run Non-negative Matrix Factorization}
\usage{
run_nmf(
  A,
  rank,
  tol = 1e-04,
  maxit = 100,
  verbose = TRUE,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  distributed_method = "Eigen"
)
}
\arguments{
\item{A}{sparse matrix giving normalized counts for genes x cells (rows x columns), or a list of sparse matrices with equal number of rows and identical rownames}

\item{rank}{factorization rank}

\item{tol}{tolerance of the fit (1e-5 for publication quality, 1e-4 for cross-validation)}

\item{maxit}{maximum number of iterations}

\item{verbose}{verbosity level}

\item{L1}{L1/LASSO penalty to increase sparsity of model}

\item{L2}{L2/Ridge penalty to increase angles between factors}

\item{threads}{number of threads for parallelization across CPUs, 0 = use all available threads}

\item{distributed_method}{experimental. Either "R" or "Eigen". For development purposes.}
}
\description{
Run NMF on a sparse matrix with automatic rank determination by cross-validation
}
