% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSharedFactors.R, R/GetUniqueFactors.R,
%   R/MetadataPlot.R, R/RunLNMF.R
\name{GetSharedFactors}
\alias{GetSharedFactors}
\alias{GetUniqueFactors}
\alias{MetadataPlot}
\alias{MetadataPlot.Seurat}
\alias{RunLNMF}
\alias{RunLNMF.Seurat}
\title{Run Linked NMF on a Seurat object}
\usage{
GetSharedFactors(object, split.by, reduction = "lnmf")

GetUniqueFactors(object, split.by, reduction = "lnmf")

MetadataPlot(object, ...)

\method{MetadataPlot}{Seurat}(object, split.by, reduction = "lnmf", ...)

RunLNMF(object, ...)

\method{RunLNMF}{Seurat}(
  object,
  split.by,
  reduction.use = "nmf",
  reduction.name = "lnmf",
  reduction.key = "LNMF_",
  verbose = TRUE,
  link.cutoff = 0.5,
  tol = 1e-05,
  maxit = 100,
  L1 = 0.01,
  L2 = 0,
  threads = 0,
  ...
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{split.by}{column name in \code{@meta.data} giving a \code{Factor} with multiple levels for splitting. Data will be weighted such that each group contributes equally to the LNMF model.}

\item{reduction}{reduction to use for metadata analysis}

\item{...}{not implemented}

\item{reduction.use}{NMF reduction to use for initializing the linked factorization.}

\item{reduction.name}{name to store resulting DimReduc object as}

\item{reduction.key}{key for resulting DimReduc}

\item{verbose}{print fitting progress to console}

\item{link.cutoff}{if the relative contribution of samples in any given group to a factor falls below \code{link.cutoff}, unlink it from the factor. \code{link.cutoff = 1} means a factor must contribute exactly equally before being unlinked.}

\item{tol}{tolerance of the fit (correlation distance of the model across consecutive iterations).}

\item{maxit}{maximum number of fitting iterations}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{L2}{L2/Ridge-like penalty to increase angles between factors}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}
}
\value{
a Seurat object with the NMF model stored in the reductions slot
}
\description{
Run a Linked Non-negative Matrix Factorization to separate shared and unique signals for integration or signature extraction.

S3 method for Seurat that runs the \code{singlet::RunLNMF} function.
}
\details{
Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunNMF}}, \code{\link{RankPlot}}, \code{\link{MetadataSummary}}
}
