% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrapper.R
\name{RunNMF.Seurat}
\alias{RunNMF.Seurat}
\alias{RunNMF}
\title{Run NMF on a Seurat object}
\usage{
\method{RunNMF}{Seurat}(
  object,
  assay = NULL,
  reduction.name = "nmf",
  reduction.key = "NMF_",
  verbose = 2,
  k = seq(2, 40, 2),
  n_replicates = 1,
  test.set.density = 0.05,
  tol = 1e-04,
  maxit = 100,
  L1 = 0.01,
  threads = 0,
  split.by = NULL,
  precision = "double",
  ...
)

RunNMF(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Assay to use, defaults to the default assay of the first object}

\item{reduction.name}{Name to store resulting DimReduc object as}

\item{reduction.key}{Key for resulting DimReduc}

\item{verbose}{Level of console output (0/FALSE, 1/TRUE, 2)}

\item{k}{vector of ranks at which to fit, witholding a test set}

\item{n_replicates}{number of replicates for cross-validation}

\item{test.set.density}{approximate density of the test set (default 0.05)}

\item{tol}{tolerance of the fit (correlation distance of the model across consecutive iterations)}

\item{maxit}{maximum number of fitting iterations}

\item{L1}{L1/LASSO penalty to increase sparsity of the model}

\item{threads}{number of threads to use (0 = let OpenMP use all available threads)}

\item{split.by}{Attribute in \code{@metadata} for splitting, if applicable. Data will be weighted such that each group contributes equally to the NMF model.}

\item{precision}{\code{"double"} or \code{"float"} for numerical precision. \code{"float"} may be faster, but numerical instability may result in more iterations to achieve desired tolerances.}

\item{...}{not implemented}
}
\value{
Returns a Seurat object with the NMF model stored in the reductions slot
}
\description{
Run a Non-negative Matrix Factorization dimensionality reduction with rank determination by cross-validation.
}
\details{
Use \code{set.seed()} to guarantee reproducibility!
}
\seealso{
\code{\link{RunLNMF}}, \code{\link{RankPlot}}, \code{\link{MetadataSummary}}
}
